+++ tests/HTTP/ClientTest.php
@@ -18,6 +18,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase {
                 CURLOPT_HEADER => true,
                 CURLOPT_POSTREDIR => 0,
                 CURLOPT_HTTPHEADER => ['X-Foo: bar'],
+                CURLOPT_NOBODY => false,
                 CURLOPT_URL => 'http://example.org/',
                 CURLOPT_CUSTOMREQUEST => 'GET',
                 CURLOPT_POSTFIELDS => null,
@@ -49,7 +50,42 @@ class ClientTest extends \PHPUnit_Framework_TestCase {
                 CURLOPT_CUSTOMREQUEST => 'HEAD',
                 CURLOPT_HTTPHEADER => ['X-Foo: bar'],
                 CURLOPT_URL => 'http://example.org/',
-                CURLOPT_POSTFIELDS => null,
+                CURLOPT_POSTFIELDS => '',
+                CURLOPT_PUT => false,
+            ];
+
+        // FIXME: CURLOPT_PROTOCOLS and CURLOPT_REDIR_PROTOCOLS are currently unsupported by HHVM
+        // at least if this unit test fails in the future we know it is :)
+        if(defined('HHVM_VERSION') === false) {
+            $settings[CURLOPT_PROTOCOLS] = CURLPROTO_HTTP | CURLPROTO_HTTPS;
+            $settings[CURLOPT_REDIR_PROTOCOLS] = CURLPROTO_HTTP | CURLPROTO_HTTPS;
+        }
+
+        $this->assertEquals($settings, $client->createCurlSettingsArray($request));
+
+    }
+
+    function testCreateCurlSettingsArrayGETAfterHEAD() {
+
+        $client = new ClientMock();
+        $request = new Request('HEAD','http://example.org/', ['X-Foo' => 'bar']);
+
+        // Parsing the settings for this method, and discarding the result.
+        // This will cause the client to automatically persist previous
+        // settings and will help us detect problems.
+        $client->createCurlSettingsArray($request);
+
+        // This is the real request.
+        $request = new Request('GET','http://example.org/', ['X-Foo' => 'bar']);
+
+        $settings = [
+                CURLOPT_CUSTOMREQUEST => 'GET',
+                CURLOPT_RETURNTRANSFER => true,
+                CURLOPT_HEADER => true,
+                CURLOPT_HTTPHEADER => ['X-Foo: bar'],
+                CURLOPT_NOBODY => false,
+                CURLOPT_URL => 'http://example.org/',
+                CURLOPT_POSTFIELDS => '',
                 CURLOPT_PUT => false,
             ];
 
@@ -77,6 +113,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase {
                 CURLOPT_HEADER => true,
                 CURLOPT_PUT => true,
                 CURLOPT_INFILE => $h,
+                CURLOPT_NOBODY => false,
                 CURLOPT_CUSTOMREQUEST => 'PUT',
                 CURLOPT_HTTPHEADER => ['X-Foo: bar'],
                 CURLOPT_URL => 'http://example.org/',
@@ -101,6 +138,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase {
         $settings = [
                 CURLOPT_RETURNTRANSFER => true,
                 CURLOPT_HEADER => true,
+                CURLOPT_NOBODY => false,
                 CURLOPT_POSTFIELDS => 'boo',
                 CURLOPT_CUSTOMREQUEST => 'PUT',
                 CURLOPT_HTTPHEADER => ['X-Foo: bar'],
@@ -337,12 +375,24 @@ class ClientTest extends \PHPUnit_Framework_TestCase {
 
 class ClientMock extends Client {
 
+    protected $persistedSettings = [];
+
     /**
      * Making this method public.
+     *
+     * We are also going to persist all settings this method generates. While
+     * the underlying object doesn't behave exactly the same, it helps us
+     * simulate what curl does internally, and helps us identify problems with
+     * settings that are set by _some_ methods and not correctly reset by other
+     * methods after subsequent use.
+     * forces
      */
     public function createCurlSettingsArray(RequestInterface $request) {
 
-        return parent::createCurlSettingsArray($request);
+        $settings = parent::createCurlSettingsArray($request);
+        $settings = $settings + $this->persistedSettings;
+        $this->persistedSettings = $settings;
+        return $settings;
 
     }
     /**