+++ CHANGELOG.md
@@ -9,6 +9,13 @@ ChangeLog
 * Added `421 Misdirected Request` from the HTTP/2.0 spec.
 
 
+3.0.5 (2015-05-11)
+------------------
+
+* #47: When re-using the client and doing any request after a `HEAD` request,
+  the client discards the body.
+
+
 3.0.4 (2014-12-10)
 ------------------